+++ lib/src/rust/test_api_route.dart
@@ -1,5 +1,3 @@
-
-
 import 'package:reduct/src/reduct_base.dart';
 
 String makeApiTest(CodeGenerator generator) {
@@ -16,14 +14,26 @@ String makeApiTest(CodeGenerator generator) {
   buffer.writeln(
     '        let app = test::init_service(App::new().app_data(data.clone()).service(get_data_api)).await;',
   );
-  buffer.writeln('        let params = QueryFilterBuilder::new().build().to_query_url();');
-  buffer.writeln('        let uri = format!("${generator.apiRoute}?{}&_limit=5", params);');
+  if (generator.onlyColumns.isEmpty) {
+    buffer.writeln(
+      '        let uri = "${generator.apiRoute}?_limit=5".to_string();',
+    );
+  } else {
+    buffer.writeln(
+      '        let params = QueryFilterBuilder::new().build().to_query_url();',
+    );
+    buffer.writeln(
+      '        let uri = format!("${generator.apiRoute}?{}&_limit=5", params);',
+    );
+  }
   buffer.writeln(
     '        let req = test::TestRequest::get().uri(&uri).to_request();',
   );
   buffer.writeln('        let resp = test::call_service(&app, req).await;');
   buffer.writeln('        assert!(resp.status().is_success());');
-  buffer.writeln('        let rs: Vec<Record> = test::read_body_json(resp).await;');
+  buffer.writeln(
+    '        let rs: Vec<Record> = test::read_body_json(resp).await;',
+  );
   buffer.writeln('        assert_eq!(rs.len(), 5);');
   buffer.writeln('    }');
   buffer.writeln('}');