+++ test/reduct_test.dart
@@ -1,7 +1,7 @@
 
 import 'rust/all_rust_test.dart' as all_rust_test;
 import 'dart/all_dart_test.dart' as all_dart_test;
-import 'utils/string_exensions_test.dart' as string_extensions_test;
+import 'utils/string_extensions_test.dart' as string_extensions_test;
 import 'html_test.dart' as html_test;
 
 void main() {